Refactoring of code
