+++ src/features/common/types/countries.d.ts
@@ -1,0 +1,5 @@
+export interface CountryType {
+  code: string;
+  label: string;
+  phone: string;
+}
